+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Use fuzzy matching with [NBT Autocomplete](https://github.com/mt1006/mc-nbtac-mod).
+
 ## [1.0.0]
 
 Initial release.